spooky sea scripts broken into two parts...
The Spooky Sea Script now is divided into two parts. When you sit on the bench
the soul will go out to sea, to return ashore as a SOULCORL, which is an item
you then give to Snorkel (dolphin) to get the SNORKEL, by standing on a dock.
New tile (which I hope to update) and "soul coral" and dolphin sprite! Timing
and other touchups.